feat: move history grouping computation to task
