+++ tools/sidebars/history-view/package.json
@@ -16,6 +16,7 @@
     "@automerge/automerge-codemirror": "^0.2.0",
     "@automerge/automerge-repo": "catalog:",
     "@automerge/automerge-repo-solid-primitives": "^2.4.0",
+    "@awarth/tasklib": "^0.0.7",
     "@inkandswitch/patchwork-elements": "workspace:^",
     "@inkandswitch/patchwork-filesystem": "workspace:^",
     "@inkandswitch/patchwork-plugins": "workspace:^",
