+++ tools/sidebars/history-view/src/highlight/useDocuments.ts
