+++ tools/sidebars/history-view/src/history/components/CopyHashButton.tsx
