+++ tools/sidebars/history-view/src/history/hooks/index.ts
@@ -1,0 +1,4 @@
+export { useCachedHistory } from "./useCachedHistory";
+export { useDocumentMetadata } from "./useDocumentMetadata";
+export { useHistorySelection } from "./useHistorySelection";
+export { useViewHeadsAnnotation } from "./useViewHeadsAnnotation";
