+++ tools/sidebars/history-view/src/history/hooks/useDocumentMetadata.ts
