--- tools/sidebars/history-view/src/utils/formatTime.ts
@@ -1,23 +1,0 @@
-import { relativeTime } from "@patchwork/util/src/relative-time";
-
-/**
- * Format a Unix timestamp (in seconds) to a display string.
- * Returns e.g. "Jan 5, 2:30 PM (3 hours ago)" or "" if no timestamp.
- */
-export function formatTime(timestampSeconds: number | undefined): string {
-  if (!timestampSeconds) return "";
-
-  const date = new Date(timestampSeconds * 1000);
-  const datePart = date.toLocaleDateString(undefined, {
-    month: "short",
-    day: "numeric",
-  });
-  const timePart = date.toLocaleTimeString(undefined, {
-    hour: "numeric",
-    minute: "2-digit",
-    hour12: true,
-  });
-  const relative = relativeTime(timestampSeconds * 1000);
-
-  return `${datePart}, ${timePart} (${relative})`;
-}
