+++ tools/sidebars/history-view/vite.config.ts
@@ -2,18 +2,26 @@ import { defineConfig } from "vite";
 import tailwindcss from "@tailwindcss/vite";
 import solid from "vite-plugin-solid";
 import cssInjectedByJsPlugin from "vite-plugin-css-injected-by-js";
-import externals from "@inkandswitch/patchwork-bootloader/externals";
+import external from "@inkandswitch/patchwork-bootloader/externals";
 
 export default defineConfig({
   base: "./",
-  plugins: [solid(), tailwindcss(), cssInjectedByJsPlugin()],
+  plugins: [
+    solid(),
+    tailwindcss(),
+    cssInjectedByJsPlugin({
+      jsAssetsFilterFunction: (chunk) => chunk.fileName === "index.js",
+    }),
+  ],
 
   build: {
-    emptyOutDir: true,
     minify: false,
     rollupOptions: {
-      external: externals,
-      input: "./src/index.tsx",
+      external,
+      input: {
+        index: "./src/index.tsx",
+        task: "./src/history/task.ts",
+      },
       output: {
         format: "es",
         entryFileNames: "[name].js",
